refactor(infra): decompose route god-file into submodules
infra/route.rs was 4230 lines (god_file) — the second-largest in the
tree, dominated by a 2515-line inline test module (bigger than its own
production code) plus a large seamless production region.

Two extractions:
- Split the mega test module into route/tests/{dispatch,handoff}.rs with
  shared test helpers (git_init, EnvGuard, CwdGuard, env_lock,
  write_rig_source_metadata, write_command_only_rig) promoted to
  route/tests/mod.rs as pub(super) so both halves reuse them.
- Extract the rig source-management cluster
  (run_rig_source_management_on_runner + path-prefix translation,
  source-sync helpers, wrapper-flag stripping) into route/rig_source.rs,
  exposed via pub(super) + use rig_source::*.

route.rs is now 1496 lines; every resulting file is under the 1500-line
god_file threshold. No logic changed; public API preserved.

Clears the infra/route.rs god_file finding.
