+++ src/commands/infra/route.rs
@@ -1245,227 +1245,8 @@ fn agent_task_fanout_cook_batch_dispatch_id(
     })
 }
 
-fn run_rig_source_management_on_runner(
-    runner_id: &str,
-    normalized_args: &[String],
-    output_file: Option<&str>,
-) -> homeboy::core::Result<(String, String, i32)> {
-    let runner = runners::load(runner_id)?;
-    let homeboy_path = runner.settings.homeboy_path.as_deref().unwrap_or("homeboy");
-    let mut command = runner_rig_source_management_command(homeboy_path, normalized_args);
-    let rig_install_source_root = rig_install_source_sync_root(&command);
-    let mut translated_remote_root = runner.workspace_root.clone().unwrap_or_default();
-    let local_cwd = std::env::current_dir().ok();
-    if let Some(local_cwd) = local_cwd.as_deref() {
-        if command_contains_path_prefix(&command, local_cwd) {
-            let (synced, sync_exit_code) = runners::sync_workspace(
-                runner_id,
-                runners::RunnerWorkspaceSyncOptions {
-                    path: local_cwd.display().to_string(),
-                    mode: runners::RunnerWorkspaceSyncMode::Snapshot,
-                    ..Default::default()
-                },
-            )?;
-            if sync_exit_code == 0 {
-                command = translate_command_path_prefix(&command, local_cwd, &synced.remote_path);
-                translated_remote_root = synced.remote_path;
-            }
-        }
-    }
-
-    // `rig install <source>` forwards the controller-local rig package path. When
-    // that package lives outside the synced working directory the runner cannot
-    // see it and fails with `Path does not exist: /Users/...`, so the documented
-    // one-command `rig install --runner` offload is broken without a local
-    // pre-install + `--skip-install` (#6964). Materialize the source's containing
-    // checkout on the runner and translate the forwarded path through the same
-    // sync+translate seam the working directory already uses. Syncing the
-    // containing checkout (not just the package directory) keeps rigs that
-    // declare `package_dependencies` — which resolve against the repo root — and
-    // package-level `extends` templates working on the runner.
-    if let Some(source_root) = rig_install_source_root.as_deref() {
-        let (synced, sync_exit_code) = runners::sync_workspace(
-            runner_id,
-            runners::RunnerWorkspaceSyncOptions {
-                path: source_root.display().to_string(),
-                mode: runners::RunnerWorkspaceSyncMode::Snapshot,
-                ..Default::default()
-            },
-        )?;
-        if sync_exit_code == 0 {
-            command = translate_command_path_prefix(&command, source_root, &synced.remote_path);
-            translated_remote_root = synced.remote_path;
-        }
-    }
-
-    command = strip_rig_source_management_local_wrapper_flags(&command);
-    if let Some(source_root) = rig_install_source_root.as_deref() {
-        runners::preflight_remote_argv_path_translation(
-            "Rig source management",
-            runner_id,
-            &command,
-            source_root,
-            &translated_remote_root,
-        )?;
-    } else if let Some(local_cwd) = local_cwd.as_deref() {
-        runners::preflight_remote_argv_path_translation(
-            "Rig source management",
-            runner_id,
-            &command,
-            local_cwd,
-            &translated_remote_root,
-        )?;
-    }
-
-    let (output, exit_code) = runners::exec(
-        runner_id,
-        RunnerExecOptions {
-            cwd: runner.workspace_root.clone(),
-            project_id: None,
-            allow_diagnostic_ssh: false,
-            diagnostic_ssh_timeout: None,
-            command,
-            env: HashMap::new(),
-            secret_env_names: Vec::new(),
-            secret_env_plan: None,
-            env_materialization: None,
-            capture_patch: false,
-            raw_exec: false,
-            source_snapshot: None,
-            path_materialization_plan: None,
-            capability_preflight: Some(runners::RunnerCapabilityPreflight {
-                command: "rig_source_management".to_string(),
-                required_commands: vec![homeboy_path.to_string()],
-                ..Default::default()
-            }),
-            required_extensions: Vec::new(),
-            accepted_extension_settings: Vec::new(),
-            require_paths: Vec::new(),
-            runner_workload: None,
-            run_id: None,
-            detach_after_handoff: false,
-            mirror_evidence: true,
-            print_handoff: true,
-        },
-    )?;
-
-    if let Some(path) = output_file {
-        write_offloaded_stdout(path, &output.stdout)?;
-    }
-
-    Ok((output.stdout, output.stderr, exit_code))
-}
-
-fn command_contains_path_prefix(command: &[String], local_root: &Path) -> bool {
-    let local_root = local_root.display().to_string();
-    let local_root = local_root.trim_end_matches('/');
-    !local_root.is_empty() && command.iter().any(|arg| arg.contains(local_root))
-}
-
-fn translate_command_path_prefix(
-    command: &[String],
-    local_root: &Path,
-    remote_root: &str,
-) -> Vec<String> {
-    let local_root = local_root.display().to_string();
-    let local_root = local_root.trim_end_matches('/');
-    let remote_root = remote_root.trim_end_matches('/');
-    command
-        .iter()
-        .map(|arg| {
-            if local_root.is_empty() || remote_root.is_empty() {
-                arg.clone()
-            } else {
-                arg.replace(local_root, remote_root)
-            }
-        })
-        .collect()
-}
-
-/// Resolve the controller-local sync root for a `rig install <source>` argument
-/// that still references a path on this controller after working-directory
-/// translation. Returns the source's containing git checkout (or the path
-/// itself when it is not inside a git repo) so the caller can materialize it on
-/// the runner and translate the forwarded argument (#6964).
-///
-/// Returns `None` for git-URL sources, paths already translated to the runner,
-/// and any source that does not exist on this controller — leaving non-path and
-/// already-remote arguments untouched.
-fn rig_install_source_sync_root(command: &[String]) -> Option<PathBuf> {
-    let source = rig_install_source_arg(command)?;
-    let expanded = shellexpand::tilde(&source).to_string();
-    let path = Path::new(&expanded);
-    if !path.exists() {
-        return None;
-    }
-    let canonical = path.canonicalize().ok()?;
-    Some(homeboy::core::git::repo_root(&canonical).unwrap_or(canonical))
-}
-
-/// Return the positional `<source>` of a `rig install` command, skipping the
-/// flags the subcommand accepts (`--id <value>`, `--id=<value>`, `--all`,
-/// `--reinstall`/`--force`). The command argv is `[homeboy, rig, install, ...]`
-/// with controller-only globals already stripped by
-/// [`runner_rig_source_management_command`].
-fn rig_install_source_arg(command: &[String]) -> Option<String> {
-    let install_index = command
-        .windows(2)
-        .position(|window| window[0] == "rig" && window[1] == "install")?
-        + 2;
-    let mut iter = command[install_index..].iter();
-    while let Some(arg) = iter.next() {
-        if arg == "--" {
-            return iter.next().cloned();
-        }
-        if arg == "--id" {
-            iter.next();
-            continue;
-        }
-        if arg.starts_with('-') {
-            continue;
-        }
-        return Some(arg.clone());
-    }
-    None
-}
-
-fn runner_rig_source_management_command(
-    homeboy_path: &str,
-    normalized_args: &[String],
-) -> Vec<String> {
-    let mut command = vec![homeboy_path.to_string()];
-    command.extend(normalized_args.iter().skip(1).cloned());
-    command
-}
-
-fn strip_rig_source_management_local_wrapper_flags(command: &[String]) -> Vec<String> {
-    const FLAGS: [(&str, bool); 7] = [
-        ("--runner", true),
-        ("--output", true),
-        ("--artifact-root", true),
-        ("--placement", true),
-        ("--allow-local-fallback", false),
-        ("--allow-dirty-lab-workspace", false),
-        ("--detach-after-handoff", false),
-    ];
-
-    let mut stripped = Vec::with_capacity(command.len());
-    let mut args = command.iter().peekable();
-    while let Some(arg) = args.next() {
-        if let Some((_, takes_value)) = FLAGS
-            .iter()
-            .find(|(flag, _)| arg == flag || arg.starts_with(&format!("{flag}=")))
-        {
-            if *takes_value && !arg.contains('=') {
-                args.next();
-            }
-            continue;
-        }
-        stripped.push(arg.clone());
-    }
-    stripped
-}
-
+mod rig_source;
+use rig_source::*;
 fn is_runs_list_runner_option(args: &[String]) -> bool {
     let Some(runs_index) = args.iter().position(|arg| arg == "runs") else {
         return false;
@@ -1712,2519 +1493,4 @@ fn strip_component_target_args(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use clap::Parser;
-    use homeboy::command_contract::{lab_runner_supports_contract_label, LabCommandPortability};
-    use std::fs;
-    use std::path::Path;
-    use std::process::Command;
-    use std::sync::{Mutex, MutexGuard, OnceLock};
-    use tempfile::tempdir;
-
-    #[test]
-    fn lab_cook_dispatcher_recipe_round_trips_exact_transport() {
-        let dispatcher = LabCookAttemptDispatcher {
-            runner_id: "homeboy-lab".to_string(),
-            allow_local_fallback: true,
-            allow_dirty_lab_workspace: false,
-            skip_deps_hydration: true,
-            source_path: Some(PathBuf::from("/controller/source")),
-            job_overrides: runners::LabJobOverrides {
-                env: [("MODE".to_string(), "test".to_string())].into(),
-                secret_env_names: vec!["TOKEN".to_string()],
-                workspace_root: Some("/runner/workspaces".to_string()),
-            },
-        };
-        let recipe =
-            crate::core::agent_task_service::AgentTaskCookAttemptDispatcher::durable_recipe(
-                &dispatcher,
-            )
-            .unwrap();
-
-        let reconstructed = reconstruct_cook_attempt_dispatcher(&recipe)
-            .unwrap()
-            .expect("Lab dispatcher reconstructed");
-
-        assert_eq!(reconstructed.durable_recipe().unwrap(), recipe);
-    }
-
-    fn git_init(path: &Path) {
-        let output = Command::new("git")
-            .args(["init", "-b", "main"])
-            .current_dir(path)
-            .output()
-            .expect("initialize git workspace");
-        assert!(
-            output.status.success(),
-            "{}",
-            String::from_utf8_lossy(&output.stderr)
-        );
-    }
-
-    struct EnvGuard {
-        previous: Vec<(&'static str, Option<String>)>,
-        _guard: MutexGuard<'static, ()>,
-    }
-
-    struct CwdGuard {
-        previous: std::path::PathBuf,
-        _guard: MutexGuard<'static, ()>,
-    }
-
-    impl EnvGuard {
-        fn set(name: &'static str, value: &str) -> Self {
-            Self::set_many(&[(name, Some(value))])
-        }
-
-        fn remove(name: &'static str) -> Self {
-            Self::set_many(&[(name, None)])
-        }
-
-        fn set_many(changes: &[(&'static str, Option<&str>)]) -> Self {
-            let guard = env_lock().lock().unwrap_or_else(|err| err.into_inner());
-            let mut previous = Vec::with_capacity(changes.len());
-            for (name, value) in changes {
-                previous.push((*name, std::env::var(name).ok()));
-                match value {
-                    Some(value) => std::env::set_var(name, value),
-                    None => std::env::remove_var(name),
-                }
-            }
-            Self {
-                previous,
-                _guard: guard,
-            }
-        }
-    }
-
-    fn env_lock() -> &'static Mutex<()> {
-        static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
-        LOCK.get_or_init(|| Mutex::new(()))
-    }
-
-    impl Drop for EnvGuard {
-        fn drop(&mut self) {
-            for (name, previous) in self.previous.iter().rev() {
-                match previous {
-                    Some(value) => std::env::set_var(name, value),
-                    None => std::env::remove_var(name),
-                }
-            }
-        }
-    }
-
-    impl CwdGuard {
-        fn set(path: &std::path::Path) -> Self {
-            let guard = env_lock().lock().unwrap_or_else(|err| err.into_inner());
-            let previous = std::env::current_dir().expect("current dir");
-            std::env::set_current_dir(path).expect("set current dir");
-            Self {
-                previous,
-                _guard: guard,
-            }
-        }
-    }
-
-    impl Drop for CwdGuard {
-        fn drop(&mut self) {
-            std::env::set_current_dir(&self.previous).expect("restore current dir");
-        }
-    }
-
-    fn write_rig_source_metadata(home: &Path, rig_id: &str, linked: bool) {
-        let sources_dir = home.join(".config").join("homeboy").join("rig-sources");
-        fs::create_dir_all(&sources_dir).expect("create rig sources dir");
-        let metadata = serde_json::json!({
-            "source": "/tmp/rig-package",
-            "source_root": "/tmp/rig-package",
-            "package_path": "/tmp/rig-package",
-            "rig_path": format!("/tmp/rig-package/rigs/{rig_id}/rig.json"),
-            "discovery_path": "/tmp/rig-package",
-            "linked": linked,
-            "materialized": false
-        });
-        fs::write(
-            sources_dir.join(format!("{rig_id}.json")),
-            serde_json::to_string_pretty(&metadata).expect("serialize rig source metadata"),
-        )
-        .expect("write rig source metadata");
-    }
-
-    fn write_command_only_rig(home: &Path, rig_id: &str) {
-        let rigs_dir = home.join(".config").join("homeboy").join("rigs");
-        fs::create_dir_all(&rigs_dir).expect("create rigs dir");
-        let spec = serde_json::json!({
-            "id": rig_id,
-            "description": "command-only rig",
-            "pipeline": {
-                "up": [
-                    {
-                        "kind": "command",
-                        "command": "./scripts/run-matrix.sh",
-                        "cwd": "tools",
-                        "env": { "MATRIX": "portable" },
-                        "label": "run matrix"
-                    }
-                ]
-            }
-        });
-        fs::write(
-            rigs_dir.join(format!("{rig_id}.json")),
-            serde_json::to_string_pretty(&spec).expect("serialize rig"),
-        )
-        .expect("write rig");
-    }
-
-    #[test]
-    fn non_lab_command_continues_local_dispatch() {
-        // route_after_parse mutates the process-global LAB_OFFLOAD_METADATA_ENV,
-        // so hold the env lock to serialize against tests that assert on it.
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        let cli = Cli::parse_from(["homeboy", "status"]);
-
-        let outcome = route_after_parse(&cli, &["homeboy".into(), "status".into()], None).unwrap();
-
-        assert_eq!(outcome, None);
-    }
-
-    #[test]
-    fn changed_scope_lint_is_lab_portable() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "review",
-            "lint",
-            "--changed-since",
-            "origin/main",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "review lint");
-        assert!(command.is_portable());
-    }
-
-    #[test]
-    fn nested_review_quality_subcommands_use_specific_lab_labels() {
-        for (args, expected_label) in [
-            (
-                vec!["homeboy", "review", "audit", "data-machine"],
-                "review audit",
-            ),
-            (
-                vec!["homeboy", "review", "lint", "data-machine"],
-                "review lint",
-            ),
-            (
-                vec!["homeboy", "review", "test", "data-machine"],
-                "review test",
-            ),
-            (
-                vec!["homeboy", "review", "build", "data-machine"],
-                "review build",
-            ),
-            (
-                vec![
-                    "homeboy",
-                    "review",
-                    "ci",
-                    "run",
-                    "data-machine",
-                    "--job",
-                    "lint",
-                ],
-                "review ci",
-            ),
-        ] {
-            let cli = Cli::parse_from(args);
-            let command = cli.command.lab_contract().unwrap();
-
-            assert_eq!(command.hot_label, expected_label);
-        }
-    }
-
-    #[test]
-    fn nested_review_lint_dispatch_uses_matching_lab_label() {
-        let cli = Cli::parse_from(["homeboy", "review", "lint", "data-machine"]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "review lint");
-        assert!(command.is_portable());
-    }
-
-    #[test]
-    fn nested_review_quality_subcommand_resolves_effective_component() {
-        let cli = Cli::parse_from(["homeboy", "review", "lint", "data-machine"]);
-        let Commands::Review(args) = cli.command else {
-            panic!("expected review command");
-        };
-
-        assert_eq!(
-            args.effective_component_args().component.as_deref(),
-            Some("data-machine")
-        );
-    }
-
-    #[test]
-    fn nested_review_quality_in_dir_offload_uses_current_dir_path() {
-        let dir = tempdir().expect("tempdir");
-        let _cwd = CwdGuard::set(dir.path());
-        let normalized = vec![
-            "homeboy".to_string(),
-            "review".to_string(),
-            "lint".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let rewritten = lab_route_source_path_args(&cli.command, &normalized, false)
-            .expect("review lint without component gets cwd path rewrite");
-        let cwd = std::env::current_dir().expect("current dir");
-
-        assert_eq!(rewritten[0..3], normalized);
-        assert_eq!(rewritten[3], "--path");
-        assert_eq!(rewritten[4], cwd.to_string_lossy());
-    }
-
-    #[test]
-    fn explicit_runner_for_changed_scope_test_is_lab_portable() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "homeboy-lab",
-            "review",
-            "test",
-            "--changed-since",
-            "origin/main",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-        assert_eq!(command.hot_label, "review test");
-        assert!(command.is_portable());
-    }
-
-    #[test]
-    fn destructive_fuzz_local_execution_requires_explicit_destructive_local_override() {
-        // Serialize against LAB_OFFLOAD_METADATA_ENV-asserting tests: this
-        // routes through route_after_parse, which mutates that global.
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        let normalized = vec![
-            "homeboy",
-            "--placement",
-            "local",
-            "fuzz",
-            "run",
-            "component-a",
-            "--allow-destructive",
-            "--isolation",
-            "isolated",
-            "--isolation-proof",
-            "proof.json",
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        assert!(destructive_fuzz_requires_lab(&cli.command));
-
-        let error = crate::test_support::with_isolated_home(|_| {
-            route_after_parse(
-                &cli,
-                &normalized
-                    .iter()
-                    .map(|arg| arg.to_string())
-                    .collect::<Vec<_>>(),
-                None,
-            )
-            .expect_err("destructive fuzz local route should be refused")
-        });
-        assert!(error
-            .to_string()
-            .contains("destructive fuzz refused local controller execution"));
-    }
-
-    #[test]
-    fn destructive_fuzz_local_override_is_command_specific_and_explicit() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--placement",
-            "local",
-            "fuzz",
-            "run",
-            "component-a",
-            "--allow-destructive",
-            "--allow-local-destructive-fuzz",
-            "--isolation",
-            "isolated",
-            "--isolation-proof",
-            "proof.json",
-        ]);
-
-        assert!(!destructive_fuzz_requires_lab(&cli.command));
-    }
-
-    #[test]
-    fn rig_up_dry_run_with_runner_emits_runner_exec_plan() {
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        crate::test_support::with_isolated_home(|home| {
-            runners::create(
-                r#"{"id":"homeboy-lab","kind":"local","homeboy_path":"/runner/bin/homeboy-patched"}"#,
-                false,
-            )
-            .expect("runner");
-            write_command_only_rig(home.path(), "script-matrix");
-            let output = home.path().join("plan.json");
-            let normalized = vec![
-                "homeboy".to_string(),
-                "--runner".to_string(),
-                "homeboy-lab".to_string(),
-                "rig".to_string(),
-                "up".to_string(),
-                "script-matrix".to_string(),
-                "--dry-run".to_string(),
-            ];
-            let cli = Cli::parse_from(&normalized);
-
-            let outcome = route_after_parse(&cli, &normalized, Some(&output.to_string_lossy()))
-                .expect("route rig up plan");
-
-            assert_eq!(outcome, Some(0));
-            let plan: serde_json::Value =
-                serde_json::from_str(&fs::read_to_string(output).expect("read output plan"))
-                    .expect("parse output plan");
-            assert_eq!(plan["variant"], "up_plan");
-            assert_eq!(plan["payload"]["runner_id"], "homeboy-lab");
-            assert_eq!(
-                plan["payload"]["selected_homeboy_binary"],
-                "/runner/bin/homeboy-patched"
-            );
-            assert_eq!(
-                plan["payload"]["commands"][0],
-                "/runner/bin/homeboy-patched runner exec homeboy-lab --cwd tools --env MATRIX=portable -- sh -c ./scripts/run-matrix.sh"
-            );
-        });
-    }
-
-    #[test]
-    fn lab_job_overrides_parse_env_json_and_workspace_root() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner-env",
-            "STUDIO_NATIVE_TRACE_SAMPLE_RUNTIME_PLUGIN_PATH=/tmp/sample-runtime",
-            "--runner-env",
-            "API_TOKEN=secret-token",
-            "--lab-env-json",
-            r#"{"EXTRA_PATH":"/tmp/extra","EMPTY":null}"#,
-            "--runner-workspace-root",
-            "/srv/job-workspace",
-            "review",
-            "test",
-            "studio-native",
-        ]);
-
-        let overrides = lab_job_overrides(&cli).expect("overrides");
-
-        assert_eq!(
-            overrides.env["STUDIO_NATIVE_TRACE_SAMPLE_RUNTIME_PLUGIN_PATH"],
-            "/tmp/sample-runtime"
-        );
-        assert_eq!(overrides.env["EXTRA_PATH"], "/tmp/extra");
-        assert_eq!(overrides.env["EMPTY"], "");
-        assert_eq!(
-            overrides.workspace_root.as_deref(),
-            Some("/srv/job-workspace")
-        );
-        assert!(overrides
-            .secret_env_names
-            .contains(&"API_TOKEN".to_string()));
-    }
-
-    #[test]
-    fn lab_job_overrides_reject_invalid_env_shapes() {
-        let cli = Cli::parse_from(["homeboy", "--runner-env", "NO_EQUALS", "review"]);
-        let err = lab_job_overrides(&cli).expect_err("invalid pair");
-        assert_eq!(err.code.as_str(), "validation.invalid_argument");
-
-        let cli = Cli::parse_from(["homeboy", "--lab-env-json", "[]", "review"]);
-        let err = lab_job_overrides(&cli).expect_err("invalid json object");
-        assert_eq!(err.code.as_str(), "validation.invalid_argument");
-    }
-
-    #[test]
-    fn changed_since_lint_keeps_git_scope_for_lab_runner() {
-        let normalized = vec![
-            "homeboy".to_string(),
-            "review".to_string(),
-            "lint".to_string(),
-            "--changed-since".to_string(),
-            "origin/main".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let rewritten = inject_lab_changed_files(&cli.command, &normalized).unwrap();
-
-        assert!(rewritten.is_none());
-    }
-
-    #[test]
-    fn changed_since_test_keeps_git_scope_for_lab_runner() {
-        let normalized = vec![
-            "homeboy".to_string(),
-            "--runner".to_string(),
-            "homeboy-lab".to_string(),
-            "review".to_string(),
-            "test".to_string(),
-            "--changed-since=origin/main".to_string(),
-            "--skip-lint".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let rewritten = inject_lab_changed_files(&cli.command, &normalized).unwrap();
-
-        assert!(rewritten.is_none());
-    }
-
-    #[test]
-    fn lab_offload_subprocess_skips_recursive_lab_routing() {
-        let _env = EnvGuard::set(
-            homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV,
-            r#"{"status":"offloaded"}"#,
-        );
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "homeboy-lab",
-            "trace",
-            "--rig",
-            "gutenberg-pattern-preview-assets",
-            "gutenberg",
-            "pattern-preview-assets",
-        ]);
-        let normalized = [
-            "homeboy".to_string(),
-            "--runner".to_string(),
-            "homeboy-lab".to_string(),
-            "trace".to_string(),
-            "--rig".to_string(),
-            "gutenberg-pattern-preview-assets".to_string(),
-            "gutenberg".to_string(),
-            "pattern-preview-assets".to_string(),
-        ];
-
-        let outcome = route_after_parse(&cli, &normalized, None).unwrap();
-
-        assert_eq!(outcome, None);
-    }
-
-    #[test]
-    fn runner_hosted_bench_exec_skips_recursive_lab_routing_without_explicit_runner() {
-        let _env = EnvGuard::set_many(&[
-            (homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV, None),
-            (homeboy::core::runner::RUNNER_HOSTED_EXEC_ENV, Some("1")),
-            (
-                homeboy::core::runner::RUNNER_PLACEMENT_RESOLVED_ENV,
-                Some("1"),
-            ),
-            (homeboy::core::runner::RUNNER_ID_ENV, Some("homeboy-lab")),
-        ]);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "--placement".to_string(),
-            "local".to_string(),
-            "bench".to_string(),
-            "--extension".to_string(),
-            "wordpress".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let outcome = route_after_parse(&cli, &normalized, None)
-            .expect("runner-hosted bench execution should stay local");
-
-        assert_eq!(outcome, None);
-    }
-
-    #[test]
-    fn ambient_resolved_marker_cannot_bypass_explicit_lab_placement() {
-        let _env = EnvGuard::set_many(&[
-            (homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV, None),
-            (homeboy::core::runner::RUNNER_HOSTED_EXEC_ENV, None),
-            (homeboy::core::runner::RUNNER_ID_ENV, None),
-            (
-                homeboy::core::runner::RUNNER_PLACEMENT_RESOLVED_ENV,
-                Some("1"),
-            ),
-        ]);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "--placement".to_string(),
-            "lab".to_string(),
-            "review".to_string(),
-            "lint".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let err = crate::test_support::with_isolated_home(|_| {
-            route_after_parse(&cli, &normalized, None)
-                .expect_err("ambient marker must not bypass required Lab placement")
-        });
-
-        assert_ne!(err.code.as_str(), "internal.unexpected");
-    }
-
-    #[test]
-    fn managed_runner_context_bypasses_auto_routing_once() {
-        let _env = EnvGuard::set_many(&[
-            (homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV, None),
-            (homeboy::core::runner::RUNNER_HOSTED_EXEC_ENV, Some("1")),
-            (
-                homeboy::core::runner::RUNNER_PLACEMENT_RESOLVED_ENV,
-                Some("1"),
-            ),
-            (homeboy::core::runner::RUNNER_ID_ENV, Some("homeboy-lab")),
-        ]);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "review".to_string(),
-            "lint".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        assert_eq!(
-            route_after_parse(&cli, &normalized, None).expect("managed context"),
-            None
-        );
-    }
-
-    #[test]
-    fn agent_task_doctor_runner_option_routes_locally() {
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "agent-task".to_string(),
-            "doctor".to_string(),
-            "--runner".to_string(),
-            "homeboy-lab".to_string(),
-            "--repair".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-
-        let outcome = route_after_parse(&cli, &normalized, None)
-            .expect("agent-task doctor owns --runner and should not be Lab-routed");
-
-        assert_eq!(outcome, None);
-        assert!(std::env::var(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV).is_err());
-    }
-
-    #[test]
-    fn trace_lab_dispatch_timeout_reads_env_override() {
-        let _env = EnvGuard::set(lab_routing::LAB_TRACE_DISPATCH_TIMEOUT_ENV, "7");
-
-        assert_eq!(
-            lab_routing::lab_trace_dispatch_timeout(),
-            std::time::Duration::from_secs(7)
-        );
-    }
-
-    #[test]
-    fn lab_route_dispatch_timeout_plumbs_core_timeout() {
-        let trace_cli = Cli::parse_from(["homeboy", "trace", "list"]);
-        let lint_cli = Cli::parse_from(["homeboy", "review", "lint"]);
-
-        assert_eq!(
-            lab_route_dispatch_timeout(&trace_cli.command),
-            Some(lab_routing::lab_trace_dispatch_timeout())
-        );
-        assert_eq!(lab_route_dispatch_timeout(&lint_cli.command), None);
-    }
-
-    #[test]
-    fn detached_agent_task_handoffs_do_not_use_trace_dispatch_timeout() {
-        let cook = Cli::parse_from([
-            "homeboy",
-            "--detach-after-handoff",
-            "agent-task",
-            "cook",
-            "--repo",
-            "homeboy",
-            "--goal",
-            "Fix the detached handoff",
-            "--to-worktree",
-            "homeboy@fix-7971",
-            "--run-id",
-            "cook-7971",
-            "--runner",
-            "homeboy-lab",
-            "--placement",
-            "lab",
-        ]);
-        let batch = Cli::parse_from([
-            "homeboy",
-            "--detach-after-handoff",
-            "agent-task",
-            "fanout",
-            "cook-batch",
-            "--repo",
-            "homeboy",
-            "--verify",
-            "cargo test --lib",
-            "--run-plan",
-            "https://github.com/Extra-Chill/homeboy/issues/7167",
-        ]);
-        let retry = Cli::parse_from([
-            "homeboy",
-            "--detach-after-handoff",
-            "agent-task",
-            "retry",
-            "failed-run",
-            "--run",
-            "--runner",
-            "homeboy-lab",
-            "--placement",
-            "lab",
-        ]);
-
-        for cli in [&cook, &batch, &retry] {
-            assert_eq!(lab_route_dispatch_timeout(&cli.command), None);
-        }
-    }
-
-    #[test]
-    fn cook_retry_lab_source_is_the_derived_baseline_not_the_controller_workspace() {
-        let baseline = tempfile::tempdir().expect("baseline");
-        let controller = tempfile::tempdir().expect("controller");
-        let capability = crate::core::agent_task_service::test_derived_cook_baseline_capability(
-            baseline.path().to_path_buf(),
-            "baseline-commit".to_string(),
-            "baseline-tree".to_string(),
-            "task",
-            Some(serde_json::json!({"workspace_snapshot_identity": "snapshot:parent"})),
-        );
-
-        assert_eq!(
-            super::cook_attempt_source_path(Some(&capability), Some(controller.path())),
-            Some(capability.canonical_path())
-        );
-        assert_eq!(
-            capability.verified_baseline_provenance(),
-            serde_json::json!({
-                "source_run_id": "test-source-run",
-                "source_task_id": "task",
-                "promoted_patch_artifact_sha256": "test-artifact-sha256",
-                "baseline_commit": "baseline-commit",
-                "baseline_tree": "baseline-tree",
-                "parent_snapshot_identity": "snapshot:parent",
-            })
-        );
-    }
-
-    #[test]
-    fn detached_retry_materializes_failed_plan_and_persists_bounded_preacceptance_failure() {
-        crate::test_support::with_isolated_home(|_| {
-            let workspace = tempfile::tempdir().expect("workspace");
-            git_init(workspace.path());
-            let source_plan = homeboy::core::agent_tasks::scheduler::AgentTaskPlan::new(
-                "failed-retry-source",
-                vec![serde_json::from_value(serde_json::json!({
-                    "task_id": "retry-task",
-                    "executor": {
-                        "backend": "fixture",
-                        "config": { "workspace_root": workspace.path() }
-                    },
-                    "instructions": "retry",
-                    "workspace": { "root": workspace.path() }
-                }))
-                .expect("task")],
-            );
-            agent_task_lifecycle::submit_plan(&source_plan, Some("failed-run"))
-                .expect("source run submitted");
-            let source_plan = agent_task_lifecycle::load_plan("failed-run").expect("source plan");
-            let failure = Error::internal_unexpected("provider exited before completion");
-            agent_task_lifecycle::record_pre_execution_failure(
-                "failed-run",
-                &source_plan,
-                "provider_execution",
-                &failure,
-            )
-            .expect("source failure persisted");
-
-            let normalized = [
-                "homeboy",
-                "--detach-after-handoff",
-                "--cwd",
-                "/controller/homeboy",
-                "agent-task",
-                "retry",
-                "failed-run",
-                "--run",
-                "--new-run-id",
-                "failed-run-retry-on-lab",
-                "--runner",
-                "homeboy-lab",
-            ]
-            .into_iter()
-            .map(str::to_string)
-            .collect::<Vec<_>>();
-            let cli = Cli::parse_from([
-                "homeboy",
-                "--detach-after-handoff",
-                "agent-task",
-                "retry",
-                "failed-run",
-                "--run",
-                "--new-run-id",
-                "failed-run-retry-on-lab",
-                "--runner",
-                "homeboy-lab",
-            ]);
-            let handoff = materialize_agent_task_retry_handoff(&cli, &normalized)
-                .expect("retry handoff materialized")
-                .expect("retry handoff");
-
-            assert!(!handoff.args.iter().any(|arg| arg == "--cwd"));
-            assert_eq!(
-                handoff.primary_workspace,
-                workspace
-                    .path()
-                    .canonicalize()
-                    .expect("canonical workspace")
-            );
-            assert!(!handoff.args.iter().any(|arg| arg == "/controller/homeboy"));
-            let agent_task_index = handoff
-                .args
-                .iter()
-                .position(|arg| arg == "agent-task")
-                .expect("agent task");
-            assert_eq!(handoff.args[agent_task_index + 1], "run-plan");
-            assert_eq!(handoff.args[agent_task_index + 2], "--plan");
-            assert_eq!(handoff.args[agent_task_index + 4], "--record-run-id");
-            assert_eq!(handoff.args[agent_task_index + 5], handoff.run_id);
-            assert_eq!(handoff.run_id, "failed-run-retry-on-lab");
-            let remote_cli = Cli::try_parse_from(&handoff.args).expect("portable run-plan argv");
-            let Commands::AgentTask(crate::commands::agent_task::AgentTaskArgs {
-                command: crate::commands::agent_task::AgentTaskCommand::RunPlan(remote),
-            }) = &remote_cli.command
-            else {
-                panic!(
-                    "Lab handoff must execute the materialized plan, not discover a retry record"
-                );
-            };
-            assert_eq!(
-                remote.record_run_id.as_deref(),
-                Some(handoff.run_id.as_str())
-            );
-            let remote_plan: homeboy::core::agent_tasks::scheduler::AgentTaskPlan =
-                serde_json::from_str(&remote.plan).expect("serialized retry plan");
-            assert_eq!(remote_plan, handoff.plan);
-            assert_eq!(
-                remote_plan.tasks[0].workspace.root.as_deref(),
-                Some(workspace.path().to_str().expect("workspace utf8"))
-            );
-            // The emitted command is accepted by the real CLI parser without
-            // inventing a global --cwd. The route carries the selected task
-            // checkout separately, and workspace staging maps it to the job cwd.
-            assert!(remote_cli.detach_after_handoff);
-            let replacement = agent_task_lifecycle::status(&handoff.run_id).expect("replacement");
-            assert_eq!(replacement.metadata["retry_of"], "failed-run");
-
-            let error = persist_retry_handoff_preacceptance_failure(
-                &handoff,
-                Error::internal_unexpected("runner preflight rejected the handoff"),
-            );
-            assert!(error
-                .hints
-                .iter()
-                .any(|hint| hint.message.contains("agent-task retry")
-                    && hint.message.contains(&handoff.run_id)));
-            let replacement = agent_task_lifecycle::status(&handoff.run_id).expect("failed retry");
-            assert_eq!(
-                replacement.state,
-                homeboy::core::agent_tasks::lifecycle::AgentTaskRunState::Failed
-            );
-            assert_eq!(
-                replacement.metadata["pre_execution_failure"]["phase"],
-                "detached_lab_handoff_preacceptance"
-            );
-        });
-    }
-
-    #[test]
-    fn controller_owned_run_materializes_plan_for_lab_execution() {
-        crate::test_support::with_isolated_home(|_| {
-            let workspace = tempfile::tempdir().expect("workspace");
-            let plan = homeboy::core::agent_tasks::scheduler::AgentTaskPlan::new(
-                "queued-controller-run",
-                vec![serde_json::from_value(serde_json::json!({
-                    "task_id": "task",
-                    "executor": { "backend": "fixture" },
-                    "instructions": "run this queued task",
-                    "workspace": { "root": workspace.path() }
-                }))
-                .expect("task")],
-            );
-            agent_task_lifecycle::submit_plan(&plan, Some("controller-queued"))
-                .expect("submit controller run");
-            let args = [
-                "homeboy",
-                "agent-task",
-                "run",
-                "controller-queued",
-                "--timeout-ms",
-                "1200",
-                "--runner",
-                "homeboy-lab",
-            ]
-            .into_iter()
-            .map(str::to_string)
-            .collect::<Vec<_>>();
-            let cli = Cli::parse_from(&args);
-
-            let handoff = materialize_agent_task_run_handoff(&cli, &args)
-                .expect("materialize run handoff")
-                .expect("controller-owned handoff");
-            let remote_cli = Cli::try_parse_from(&handoff.args).expect("portable run-plan argv");
-            let Commands::AgentTask(crate::commands::agent_task::AgentTaskArgs {
-                command: crate::commands::agent_task::AgentTaskCommand::RunPlan(remote),
-            }) = remote_cli.command
-            else {
-                panic!("controller run must execute its materialized plan on Lab");
-            };
-
-            assert_eq!(remote.record_run_id.as_deref(), Some("controller-queued"));
-            assert_eq!(remote.timeout_ms, Some(1200));
-            assert_eq!(
-                serde_json::from_str::<homeboy::core::agent_tasks::scheduler::AgentTaskPlan>(
-                    &remote.plan
-                )
-                .expect("serialized plan"),
-                plan
-            );
-            assert_eq!(
-                handoff.primary_workspace,
-                workspace.path().canonicalize().unwrap()
-            );
-        });
-    }
-
-    #[test]
-    fn controller_owned_run_refuses_to_handoff_a_plan_without_a_workspace() {
-        crate::test_support::with_isolated_home(|_| {
-            let plan = homeboy::core::agent_tasks::scheduler::AgentTaskPlan::new(
-                "queued-controller-run-without-workspace",
-                vec![serde_json::from_value(serde_json::json!({
-                    "task_id": "task",
-                    "executor": { "backend": "fixture" },
-                    "instructions": "run this queued task"
-                }))
-                .expect("task")],
-            );
-            agent_task_lifecycle::submit_plan(&plan, Some("controller-queued-without-workspace"))
-                .expect("submit controller run");
-            let args = [
-                "homeboy",
-                "agent-task",
-                "run",
-                "controller-queued-without-workspace",
-                "--runner",
-                "homeboy-lab",
-            ]
-            .into_iter()
-            .map(str::to_string)
-            .collect::<Vec<_>>();
-            let cli = Cli::parse_from(&args);
-
-            let error = materialize_agent_task_run_handoff(&cli, &args)
-                .expect_err("workspace-less controller plan must fail before Lab handoff");
-
-            assert_eq!(error.details["field"], "workspace");
-            assert!(error
-                .message
-                .contains("requires exactly one task workspace"));
-        });
-    }
-
-    #[test]
-    fn controller_owned_run_refuses_an_ambiguous_plan_workspace() {
-        let first = tempfile::tempdir().expect("first workspace");
-        let second = tempfile::tempdir().expect("second workspace");
-        let plan = homeboy::core::agent_tasks::scheduler::AgentTaskPlan::new(
-            "ambiguous-controller-run",
-            vec![
-                serde_json::from_value(serde_json::json!({
-                    "task_id": "first",
-                    "executor": { "backend": "fixture" },
-                    "instructions": "first task",
-                    "workspace": { "root": first.path() }
-                }))
-                .expect("first task"),
-                serde_json::from_value(serde_json::json!({
-                    "task_id": "second",
-                    "executor": { "backend": "fixture" },
-                    "instructions": "second task",
-                    "workspace": { "root": second.path() }
-                }))
-                .expect("second task"),
-            ],
-        );
-
-        let error = plan_primary_workspace(&plan)
-            .expect_err("multiple controller plan workspaces must fail before handoff");
-
-        assert_eq!(error.details["field"], "workspace");
-        assert!(error.message.contains("multiple task workspaces"));
-    }
-
-    #[test]
-    fn lab_owned_retry_is_not_read_from_the_controller_store() {
-        crate::test_support::with_isolated_home(|_| {
-            let args = [
-                "homeboy",
-                "agent-task",
-                "retry",
-                "lab-owned-failed-run",
-                "--run",
-                "--runner",
-                "homeboy-lab",
-            ]
-            .into_iter()
-            .map(str::to_string)
-            .collect::<Vec<_>>();
-            let cli = Cli::parse_from(&args);
-
-            assert!(materialize_agent_task_retry_handoff(&cli, &args)
-                .expect("runner-owned retry stays portable")
-                .is_none());
-        });
-    }
-
-    #[test]
-    fn explicit_local_cook_does_not_enter_lab_attempt_dispatch() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--placement",
-            "local",
-            "agent-task",
-            "cook",
-            "--to-worktree",
-            "fixture@local",
-            "--verify",
-            "true",
-            "--prompt",
-            "keep this local",
-        ]);
-
-        assert!(
-            run_split_placement_cook(&cli, &[], None, Some("homeboy-lab"))
-                .expect("local placement bypasses Lab cook dispatch")
-                .is_none()
-        );
-    }
-
-    #[test]
-    fn cook_to_worktree_provider_workspace_survives_failed_attempt_and_lab_retry() {
-        crate::test_support::with_isolated_home(|_| {
-            let workspace = tempfile::tempdir().expect("workspace");
-            git_init(workspace.path());
-            let provider_dir = tempfile::tempdir().expect("provider dir");
-            let provider = provider_dir.path().join("provider");
-            let payload = serde_json::json!({
-                "worktrees": [{
-                    "handle": "blocks-engine@cook-six-fixture-generic-parity",
-                    "path": workspace.path(),
-                    "branch": "main",
-                    "safety": { "dirty": false, "unpushed": false, "primary": false }
-                }]
-            });
-            fs::write(
-                &provider,
-                format!("#!/bin/sh\nprintf '%s\\n' '{}'\n", payload),
-            )
-            .expect("write provider");
-            #[cfg(unix)]
-            {
-                use std::os::unix::fs::PermissionsExt;
-
-                let mut permissions = fs::metadata(&provider)
-                    .expect("provider metadata")
-                    .permissions();
-                permissions.set_mode(0o755);
-                fs::set_permissions(&provider, permissions).expect("make provider executable");
-            }
-            let mut config = homeboy::core::defaults::HomeboyConfig::default();
-            config.worktree_providers.insert(
-                "fixture".to_string(),
-                homeboy::core::defaults::WorktreeProviderConfig {
-                    enabled: true,
-                    kind: homeboy::core::defaults::WorktreeProviderKind::Command,
-                    apply_enabled: false,
-                    commands: homeboy::core::defaults::WorktreeProviderCommands {
-                        resolve: Some(vec![provider.display().to_string(), "{handle}".to_string()]),
-                        ..Default::default()
-                    },
-                    list_result_mapping: Some(
-                        homeboy::core::defaults::WorktreeProviderListResultMapping {
-                            items: "$.worktrees".to_string(),
-                            handle: "$.handle".to_string(),
-                            path: "$.path".to_string(),
-                            branch: "$.branch".to_string(),
-                            dirty: "$.safety.dirty".to_string(),
-                            unpushed: "$.safety.unpushed".to_string(),
-                            primary: "$.safety.primary".to_string(),
-                        },
-                    ),
-                },
-            );
-            homeboy::core::defaults::save_config(&config).expect("save provider config");
-
-            let cook_cli = Cli::parse_from([
-                "homeboy",
-                "agent-task",
-                "cook",
-                "--to-worktree",
-                "blocks-engine@cook-six-fixture-generic-parity",
-                "--verify",
-                "true",
-                "--backend",
-                "fixture",
-                "--prompt",
-                "retry this task",
-            ]);
-            let plan = materialize_agent_task_cook_plan(&cook_cli)
-                .expect("materialize cook plan")
-                .expect("cook plan");
-            let expected_root = workspace.path().display().to_string();
-            assert_eq!(
-                plan.tasks[0].workspace.root.as_deref(),
-                Some(expected_root.as_str())
-            );
-            agent_task_lifecycle::submit_plan(&plan, Some("failed-run")).expect("submit plan");
-            agent_task_lifecycle::record_pre_execution_failure(
-                "failed-run",
-                &plan,
-                "lab_handoff_preacceptance",
-                &Error::internal_unexpected("Lab rejected the initial attempt"),
-            )
-            .expect("persist failed attempt");
-
-            let retry_args = ["homeboy", "agent-task", "retry", "failed-run", "--run"]
-                .into_iter()
-                .map(str::to_string)
-                .collect::<Vec<_>>();
-            let retry_cli = Cli::parse_from(&retry_args);
-            let handoff = materialize_agent_task_retry_handoff(&retry_cli, &retry_args)
-                .expect("materialize retry handoff")
-                .expect("retry handoff");
-
-            assert_eq!(
-                handoff.primary_workspace,
-                workspace.path().canonicalize().expect("root")
-            );
-            assert_eq!(
-                handoff.plan.tasks[0].workspace.root.as_deref(),
-                Some(expected_root.as_str())
-            );
-        });
-    }
-
-    #[test]
-    fn retry_handoff_refuses_multiple_task_workspaces() {
-        crate::test_support::with_isolated_home(|_| {
-            let first = tempfile::tempdir().expect("first workspace");
-            let second = tempfile::tempdir().expect("second workspace");
-            git_init(first.path());
-            git_init(second.path());
-            let task = |id: &str, root: &Path| {
-                serde_json::from_value(serde_json::json!({
-                    "task_id": id,
-                    "executor": { "backend": "fixture" },
-                    "instructions": "retry",
-                    "workspace": { "root": root }
-                }))
-                .expect("task")
-            };
-            let plan = homeboy::core::agent_tasks::scheduler::AgentTaskPlan::new(
-                "multiple-workspaces",
-                vec![task("first", first.path()), task("second", second.path())],
-            );
-            agent_task_lifecycle::submit_plan(&plan, Some("failed-run")).expect("source plan");
-            let source_plan = agent_task_lifecycle::load_plan("failed-run").expect("source plan");
-            agent_task_lifecycle::record_pre_execution_failure(
-                "failed-run",
-                &source_plan,
-                "provider_execution",
-                &Error::internal_unexpected("failed"),
-            )
-            .expect("source failure");
-            let normalized = ["homeboy", "agent-task", "retry", "failed-run", "--run"]
-                .into_iter()
-                .map(str::to_string)
-                .collect::<Vec<_>>();
-            let cli = Cli::parse_from(&normalized);
-
-            let error = match materialize_agent_task_retry_handoff(&cli, &normalized) {
-                Ok(_) => panic!("multiple workspaces must fail closed"),
-                Err(error) => error,
-            };
-            assert!(error.message.contains("multiple task workspaces"));
-            assert!(agent_task_lifecycle::status("failed-run-retry-1").is_err());
-        });
-    }
-
-    #[test]
-    fn retry_handoff_identifies_an_original_plan_without_a_workspace() {
-        crate::test_support::with_isolated_home(|_| {
-            let plan = homeboy::core::agent_tasks::scheduler::AgentTaskPlan::new(
-                "missing-workspace",
-                vec![serde_json::from_value(serde_json::json!({
-                    "task_id": "retry-task",
-                    "executor": { "backend": "fixture" },
-                    "instructions": "retry"
-                }))
-                .expect("task")],
-            );
-            agent_task_lifecycle::submit_plan(&plan, Some("failed-run")).expect("source plan");
-            let source_plan = agent_task_lifecycle::load_plan("failed-run").expect("source plan");
-            agent_task_lifecycle::record_pre_execution_failure(
-                "failed-run",
-                &source_plan,
-                "provider_execution",
-                &Error::internal_unexpected("failed"),
-            )
-            .expect("source failure");
-            let normalized = ["homeboy", "agent-task", "retry", "failed-run", "--run"]
-                .into_iter()
-                .map(str::to_string)
-                .collect::<Vec<_>>();
-            let cli = Cli::parse_from(&normalized);
-
-            let error = match materialize_agent_task_retry_handoff(&cli, &normalized) {
-                Ok(_) => panic!("missing original workspace must fail before creating a retry"),
-                Err(error) => error,
-            };
-
-            assert!(error.message.contains("original persisted plan has none"));
-            assert!(agent_task_lifecycle::status("failed-run-retry-1").is_err());
-        });
-    }
-
-    #[test]
-    fn agent_task_fanout_dispatch_id_uses_explicit_or_stable_default() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--detach-after-handoff",
-            "agent-task",
-            "fanout",
-            "cook-batch",
-            "--repo",
-            "homeboy",
-            "--fanout-id",
-            "wave-7167",
-            "--verify",
-            "cargo test --lib",
-            "--run-plan",
-            "https://github.com/Extra-Chill/homeboy/issues/7167",
-        ]);
-        let Commands::AgentTask(crate::commands::agent_task::AgentTaskArgs {
-            command:
-                crate::commands::agent_task::AgentTaskCommand::Fanout(
-                    crate::commands::agent_task::AgentTaskFanoutArgs {
-                        command:
-                            crate::commands::agent_task::AgentTaskFanoutCommand::CookBatch(args),
-                    },
-                ),
-        }) = cli.command
-        else {
-            panic!("cook-batch command");
-        };
-
-        assert_eq!(agent_task_fanout_cook_batch_dispatch_id(&args), "wave-7167");
-
-        let mut default_args = args;
-        default_args.fanout_id = None;
-        assert_eq!(
-            agent_task_fanout_cook_batch_dispatch_id(&default_args),
-            "cook-batch-homeboy-issue-7167-1"
-        );
-    }
-
-    #[test]
-    fn agent_task_fanout_finish_metadata_preserves_discoverability_commands() {
-        let metadata = agent_task_fanout_finish_metadata(
-            serde_json::json!({
-                "lab_dispatch": {
-                    "status": "error",
-                    "runner_id": "homeboy-lab",
-                },
-            }),
-            "dispatch-run-7167",
-            "cook-batch-homeboy-issue-7167-1",
-            RunStatus::Error,
-        );
-
-        assert_eq!(
-            metadata["agent_task_lab_dispatch"]["fanout_id"],
-            "cook-batch-homeboy-issue-7167-1"
-        );
-        assert_eq!(metadata["agent_task_lab_dispatch"]["status"], "error");
-        assert_eq!(
-            metadata["follow_commands"]["dispatch_status"],
-            "homeboy runs show dispatch-run-7167"
-        );
-        assert_eq!(
-            metadata["follow_commands"]["dispatch_evidence"],
-            "homeboy runs evidence --run dispatch-run-7167"
-        );
-        assert_eq!(
-            metadata["follow_commands"]["fanout_status"],
-            "homeboy agent-task fanout status cook-batch-homeboy-issue-7167-1"
-        );
-    }
-
-    #[test]
-    fn offloaded_stdout_write_preserves_bytes_for_output_file() {
-        let dir = tempdir().unwrap();
-        let output_path = dir.path().join("out.json");
-
-        write_offloaded_stdout(&output_path.to_string_lossy(), "{\"ok\":true}\n").unwrap();
-
-        assert_eq!(
-            std::fs::read_to_string(output_path).unwrap(),
-            "{\"ok\":true}\n"
-        );
-    }
-
-    #[test]
-    fn runner_rig_source_management_remote_preflight_strips_controller_globals() {
-        let normalized = vec![
-            "homeboy".to_string(),
-            "rig".to_string(),
-            "sources".to_string(),
-            "list".to_string(),
-            "--runner".to_string(),
-            "homeboy-lab".to_string(),
-            "--output=./sources.json".to_string(),
-            "--placement".to_string(),
-            "lab-or-local".to_string(),
-            "--placement=lab".to_string(),
-            "--detach-after-handoff".to_string(),
-        ];
-
-        let command = runner_rig_source_management_command("/usr/local/bin/homeboy", &normalized);
-        let preflight = strip_rig_source_management_local_wrapper_flags(&command);
-
-        assert_eq!(
-            preflight,
-            vec![
-                "/usr/local/bin/homeboy".to_string(),
-                "rig".to_string(),
-                "sources".to_string(),
-                "list".to_string(),
-            ]
-        );
-    }
-
-    #[test]
-    fn runner_rig_source_management_translates_local_subdir_paths() {
-        let command = vec![
-            "/runner/bin/homeboy".to_string(),
-            "rig".to_string(),
-            "install".to_string(),
-            "/Users/chubes/Developer/homeboy-rigs@run/WordPress/static-site-importer".to_string(),
-        ];
-
-        let translated = translate_command_path_prefix(
-            &command,
-            std::path::Path::new("/Users/chubes/Developer/homeboy-rigs@run"),
-            "/home/chubes/Developer/_lab_workspaces/homeboy-rigs-run-abc",
-        );
-
-        assert_eq!(
-            translated[3],
-            "/home/chubes/Developer/_lab_workspaces/homeboy-rigs-run-abc/WordPress/static-site-importer"
-        );
-    }
-
-    #[test]
-    fn rig_install_source_arg_finds_positional_source_after_flags() {
-        let command = vec![
-            "/runner/bin/homeboy".to_string(),
-            "rig".to_string(),
-            "install".to_string(),
-            "--id".to_string(),
-            "static-site-importer".to_string(),
-            "--reinstall".to_string(),
-            "/Users/chubes/Developer/homeboy-rigs@run/WordPress/static-site-importer".to_string(),
-            "--all".to_string(),
-        ];
-
-        assert_eq!(
-            rig_install_source_arg(&command).as_deref(),
-            Some("/Users/chubes/Developer/homeboy-rigs@run/WordPress/static-site-importer")
-        );
-    }
-
-    #[test]
-    fn rig_install_source_arg_ignores_non_install_commands() {
-        let command = vec![
-            "/runner/bin/homeboy".to_string(),
-            "rig".to_string(),
-            "sources".to_string(),
-            "list".to_string(),
-        ];
-
-        assert_eq!(rig_install_source_arg(&command), None);
-    }
-
-    #[test]
-    fn rig_install_source_sync_root_resolves_existing_local_package() {
-        let source_dir = tempdir().expect("source dir");
-        let source_path = source_dir
-            .path()
-            .canonicalize()
-            .expect("canonical temp dir")
-            .join("static-site-importer");
-        fs::create_dir_all(&source_path).expect("create source package");
-        let command = vec![
-            "/runner/bin/homeboy".to_string(),
-            "rig".to_string(),
-            "install".to_string(),
-            source_path.to_string_lossy().to_string(),
-        ];
-
-        let sync_root = rig_install_source_sync_root(&command).expect("sync root");
-
-        // The temp dir is not a git repo, so the package directory itself is the
-        // materialization root.
-        assert_eq!(sync_root, source_path);
-    }
-
-    #[test]
-    fn rig_install_source_sync_root_skips_git_url_and_missing_paths() {
-        let git_url = vec![
-            "/runner/bin/homeboy".to_string(),
-            "rig".to_string(),
-            "install".to_string(),
-            "https://github.com/Extra-Chill/homeboy-rigs.git".to_string(),
-        ];
-        assert_eq!(rig_install_source_sync_root(&git_url), None);
-
-        let missing = vec![
-            "/runner/bin/homeboy".to_string(),
-            "rig".to_string(),
-            "install".to_string(),
-            "/Users/chubes/Developer/does-not-exist-rig-package-6964".to_string(),
-        ];
-        assert_eq!(rig_install_source_sync_root(&missing), None);
-    }
-
-    #[test]
-    fn rig_install_offload_translates_source_path_instead_of_forwarding_it() {
-        let source_dir = tempdir().expect("source dir");
-        let source_path = source_dir
-            .path()
-            .canonicalize()
-            .expect("canonical temp dir")
-            .join("static-site-importer");
-        fs::create_dir_all(&source_path).expect("create source package");
-        let local_source = source_path.to_string_lossy().to_string();
-        let command = vec![
-            "/runner/bin/homeboy".to_string(),
-            "rig".to_string(),
-            "install".to_string(),
-            local_source.clone(),
-            "--reinstall".to_string(),
-        ];
-
-        let sync_root = rig_install_source_sync_root(&command).expect("sync root");
-        let remote_root = "/home/runner/Developer/_lab_workspaces/static-site-importer-abc";
-        let translated = translate_command_path_prefix(&command, &sync_root, remote_root);
-
-        // The forwarded source must be the runner-side path, never the
-        // controller-local path that broke `rig install --runner` (#6964).
-        assert_eq!(translated[3], remote_root);
-        assert!(
-            !translated.iter().any(|arg| arg.contains(&local_source)),
-            "controller-local source path must not be forwarded: {translated:?}"
-        );
-    }
-
-    #[test]
-    fn linked_local_rig_check_disables_default_lab_offload() {
-        let temp_home = tempdir().expect("temp home");
-        let _home = EnvGuard::set("HOME", temp_home.path().to_str().expect("home path"));
-        write_rig_source_metadata(temp_home.path(), "linked-local", true);
-        let cli = Cli::parse_from(["homeboy", "rig", "check", "linked-local"]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "rig check");
-        assert!(command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-        assert!(!command.routing_policy.infer_source_path_tools);
-        assert!(cli.command.supports_lab_runner());
-    }
-
-    #[test]
-    fn linked_local_rig_check_stays_local_without_runner() {
-        // Scope the offload-metadata env var so a parallel test that sets it
-        // (process-global) cannot leak into this local/no-runner assertion.
-        let temp_home = tempdir().expect("temp home");
-        let _env = EnvGuard::set_many(&[
-            (homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV, None),
-            ("HOME", Some(temp_home.path().to_str().expect("home path"))),
-        ]);
-        write_rig_source_metadata(temp_home.path(), "linked-local", true);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "rig".to_string(),
-            "check".to_string(),
-            "linked-local".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let outcome = route_after_parse(&cli, &normalized, None)
-            .expect("linked local rig check should skip automatic Lab offload");
-
-        assert_eq!(outcome, None);
-        assert!(std::env::var(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV).is_err());
-    }
-
-    #[test]
-    fn installed_git_rig_check_keeps_default_lab_offload() {
-        let temp_home = tempdir().expect("temp home");
-        let _home = EnvGuard::set("HOME", temp_home.path().to_str().expect("home path"));
-        write_rig_source_metadata(temp_home.path(), "installed-git", false);
-        let cli = Cli::parse_from(["homeboy", "rig", "check", "installed-git"]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "rig check");
-        assert!(command.is_portable());
-        assert!(command.routing_policy.default_lab_offload);
-        assert!(!command.routing_policy.infer_source_path_tools);
-    }
-
-    #[test]
-    fn lab_command_preserves_portable_contract_shape() {
-        let cli = Cli::parse_from(["homeboy", "review", "lint"]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "review lint");
-        assert!(command.is_portable());
-        assert!(command.routing_policy.requires_extension_parity);
-    }
-
-    #[test]
-    fn extension_update_requires_explicit_lab_runner_without_extension_parity() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "lab",
-            "extension",
-            "update",
-            "wordpress",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "extension update");
-        assert!(command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-        assert!(!command.routing_policy.requires_extension_parity);
-        assert!(command.required_extensions.is_empty());
-        assert!(!command.routing_policy.infer_source_path_tools);
-        assert!(cli.command.supports_lab_runner());
-    }
-
-    #[test]
-    fn extension_refresh_requires_explicit_lab_runner_without_extension_parity() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "lab",
-            "extension",
-            "refresh",
-            "https://github.com/Extra-Chill/homeboy-extensions.git",
-            "--id",
-            "wordpress",
-            "--ref",
-            "6ff93f43",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "extension refresh");
-        assert!(command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-        assert!(!command.routing_policy.requires_extension_parity);
-        assert!(command.required_extensions.is_empty());
-        assert!(!command.routing_policy.infer_source_path_tools);
-        assert!(cli.command.supports_lab_runner());
-    }
-
-    #[test]
-    fn extension_update_routes_locally_without_explicit_lab_runner() {
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "extension".to_string(),
-            "update".to_string(),
-            "wordpress".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let outcome = route_after_parse(&cli, &normalized, None)
-            .expect("extension update without --runner should not offload");
-
-        assert_eq!(outcome, None);
-        assert!(std::env::var(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV).is_err());
-    }
-
-    #[test]
-    fn extension_dev_run_keeps_its_runner_workflow_on_the_controller() {
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "extension".to_string(),
-            "dev-run".to_string(),
-            "wordpress".to_string(),
-            "--source".to_string(),
-            "/tmp/wordpress".to_string(),
-            "--runner".to_string(),
-            "homeboy-lab".to_string(),
-            "homeboy".to_string(),
-            "extension".to_string(),
-            "show".to_string(),
-            "wordpress".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let outcome = route_after_parse(&cli, &normalized, None)
-            .expect("dev-run should execute its own runner lifecycle");
-
-        assert_eq!(outcome, None);
-        assert!(std::env::var(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV).is_err());
-    }
-
-    #[test]
-    fn extension_show_routes_to_explicit_lab_runner() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "lab",
-            "extension",
-            "show",
-            "wordpress",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "extension show");
-        assert!(command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-        assert!(!command.routing_policy.requires_extension_parity);
-        assert!(command.required_extensions.is_empty());
-        assert!(!command.routing_policy.infer_source_path_tools);
-        assert!(cli.command.supports_lab_runner());
-    }
-
-    #[test]
-    fn fuzz_doctor_supports_runner_lab_placement_route() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "fuzz",
-            "doctor",
-            "--extension",
-            "nodejs",
-            "--runner",
-            "homeboy-lab",
-            "--placement",
-            "lab",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-        assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-        assert_eq!(cli.placement, crate::cli_surface::Placement::Lab);
-        assert_eq!(command.hot_label, "fuzz doctor");
-        assert!(lab_runner_supports_contract_label(command.hot_label));
-        assert!(command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-        assert!(command.routing_policy.requires_extension_parity);
-        assert!(command.routing_policy.read_only_polling);
-        assert_eq!(command.required_extensions, vec!["nodejs".to_string()]);
-        assert_eq!(
-            command.source_path_mode,
-            runners::LabOffloadSourcePathMode::RunnerResident
-        );
-        assert_eq!(
-            command.workspace_mode_policy,
-            runners::LabOffloadWorkspaceModePolicy::RunnerResident
-        );
-        assert!(cli.command.lab_offload_mutation_flag().is_none());
-        assert!(!cli.command.lab_offload_captures_mutation_patch());
-        assert!(cli.command.supports_lab_runner());
-    }
-
-    #[test]
-    fn fuzz_doctor_routes_locally_without_explicit_lab_runner() {
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "fuzz".to_string(),
-            "doctor".to_string(),
-            "--extension".to_string(),
-            "nodejs".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let outcome = route_after_parse(&cli, &normalized, None)
-            .expect("fuzz doctor without --runner should remain a local diagnostic");
-
-        assert_eq!(outcome, None);
-        assert!(std::env::var(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV).is_err());
-    }
-
-    #[test]
-    fn extension_list_stays_local_only() {
-        let cli = Cli::parse_from(["homeboy", "--runner", "lab", "extension", "list"]);
-
-        assert!(lab_offload_command(&cli.command).unwrap().is_none());
-        assert!(!cli.command.supports_lab_runner());
-    }
-
-    #[test]
-    fn global_runner_for_runs_show_has_local_mirror_guidance() {
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "homeboy-lab",
-            "runs",
-            "show",
-            "run-123",
-        ]);
-
-        let err = route_after_parse(
-            &cli,
-            &[
-                "homeboy".into(),
-                "--runner".into(),
-                "homeboy-lab".into(),
-                "runs".into(),
-                "show".into(),
-                "run-123".into(),
-            ],
-            None,
-        )
-        .expect_err("runs show rejects global runner with guidance");
-
-        assert_eq!(err.code.as_str(), "validation.invalid_argument");
-        assert!(err.message.contains("homeboy runs show run-123"));
-        assert!(err.message.contains("without --runner"));
-    }
-
-    #[test]
-    fn runs_list_runner_option_after_subcommand_routes_locally() {
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-
-        for normalized in [
-            vec![
-                "homeboy".to_string(),
-                "runs".to_string(),
-                "list".to_string(),
-                "--runner".to_string(),
-                "homeboy-lab".to_string(),
-                "--status".to_string(),
-                "running".to_string(),
-                "--limit".to_string(),
-                "20".to_string(),
-            ],
-            vec![
-                "homeboy".to_string(),
-                "runs".to_string(),
-                "list".to_string(),
-                "--runner=homeboy-lab".to_string(),
-                "--status".to_string(),
-                "running".to_string(),
-                "--limit".to_string(),
-                "20".to_string(),
-            ],
-        ] {
-            let cli = Cli::parse_from(&normalized);
-
-            let outcome = route_after_parse(&cli, &normalized, None)
-                .expect("runs list subcommand runner option should not be rejected");
-
-            assert_eq!(outcome, None);
-        }
-    }
-
-    #[test]
-    fn global_runner_for_runs_list_keeps_placement_guidance() {
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "--runner".to_string(),
-            "homeboy-lab".to_string(),
-            "runs".to_string(),
-            "list".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let err = route_after_parse(&cli, &normalized, None)
-            .expect_err("top-level runner on runs list should keep placement guidance");
-
-        assert_eq!(err.code.as_str(), "validation.invalid_argument");
-        assert!(err
-            .message
-            .contains("homeboy runs list --runner homeboy-lab"));
-    }
-
-    #[test]
-    fn runs_artifact_attach_runner_option_routes_locally() {
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-
-        for normalized in [
-            vec![
-                "homeboy".to_string(),
-                "runs".to_string(),
-                "artifact".to_string(),
-                "attach".to_string(),
-                "--runner".to_string(),
-                "homeboy-lab".to_string(),
-                "--path".to_string(),
-                "/tmp/matrix-summary.json".to_string(),
-                "--name".to_string(),
-                "matrix-summary".to_string(),
-                "run-123".to_string(),
-            ],
-            vec![
-                "homeboy".to_string(),
-                "runs".to_string(),
-                "artifact".to_string(),
-                "attach".to_string(),
-                "--runner=homeboy-lab".to_string(),
-                "--path=/tmp/matrix-summary.json".to_string(),
-                "--name=matrix-summary".to_string(),
-                "run-123".to_string(),
-            ],
-        ] {
-            let cli = Cli::parse_from(&normalized);
-
-            let outcome = route_after_parse(&cli, &normalized, None)
-                .expect("runs artifact attach command-local runner option should not be rejected");
-
-            assert_eq!(outcome, None);
-        }
-    }
-
-    #[test]
-    fn agent_task_inspection_commands_support_runner_resident_recovery() {
-        for args in [
-            ["homeboy", "agent-task", "status", "agent-task-123"].as_slice(),
-            ["homeboy", "agent-task", "logs", "agent-task-123"].as_slice(),
-            ["homeboy", "agent-task", "artifacts", "agent-task-123"].as_slice(),
-            ["homeboy", "agent-task", "review", "agent-task-123"].as_slice(),
-        ] {
-            let cli = Cli::parse_from(args);
-            let command = lab_offload_command(&cli.command).unwrap().unwrap();
-            assert!(lab_runner_supports_contract_label(command.hot_label));
-            assert_eq!(
-                command.source_path_mode,
-                runners::LabOffloadSourcePathMode::RunnerResident
-            );
-            assert_eq!(
-                command.workspace_mode_policy,
-                runners::LabOffloadWorkspaceModePolicy::RunnerResident
-            );
-            assert!(!command.routing_policy.default_lab_offload);
-        }
-    }
-
-    #[test]
-    fn agent_task_retry_run_supports_explicit_runner() {
-        for args in [
-            [
-                "homeboy",
-                "--runner",
-                "homeboy-lab",
-                "agent-task",
-                "retry",
-                "agent-task-123",
-                "--run",
-            ],
-            [
-                "homeboy",
-                "agent-task",
-                "retry",
-                "agent-task-123",
-                "--run",
-                "--runner",
-                "homeboy-lab",
-            ],
-        ] {
-            let cli = Cli::parse_from(args);
-
-            let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-            assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-            assert!(lab_runner_supports_contract_label(command.hot_label));
-            assert!(command.is_portable());
-            assert!(command.routing_policy.default_lab_offload);
-        }
-    }
-
-    #[test]
-    fn agent_task_cook_keeps_its_coordinator_local_for_all_placements() {
-        let automatic = Cli::parse_from([
-            "homeboy",
-            "agent-task",
-            "cook",
-            "--prompt",
-            "implement the fix",
-            "--to-worktree",
-            "homeboy@cook-routing",
-            "--verify",
-            "cargo test --locked",
-        ]);
-        let automatic_command = lab_offload_command(&automatic.command).unwrap().unwrap();
-        assert!(!automatic_command.is_portable());
-        assert!(!automatic_command.routing_policy.default_lab_offload);
-
-        let explicit = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "homeboy-lab",
-            "agent-task",
-            "cook",
-            "--prompt",
-            "implement the fix",
-            "--to-worktree",
-            "homeboy@cook-routing",
-            "--verify",
-            "cargo test --locked",
-        ]);
-        let explicit_command = lab_offload_command(&explicit.command).unwrap().unwrap();
-        assert_eq!(explicit.runner.as_deref(), Some("homeboy-lab"));
-        assert!(!explicit_command.is_portable());
-
-        let lab = Cli::parse_from([
-            "homeboy",
-            "--placement",
-            "lab",
-            "agent-task",
-            "cook",
-            "--prompt",
-            "implement the fix",
-            "--to-worktree",
-            "homeboy@cook-routing",
-            "--verify",
-            "cargo test --locked",
-        ]);
-        assert!(!lab_offload_command(&lab.command)
-            .unwrap()
-            .unwrap()
-            .is_portable());
-        assert_eq!(lab.placement, crate::cli_surface::Placement::Lab);
-    }
-
-    #[test]
-    fn agent_task_providers_supports_explicit_runner_discovery() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "homeboy-lab",
-            "agent-task",
-            "providers",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-        assert!(lab_runner_supports_contract_label(command.hot_label));
-        assert!(command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-        assert!(!command.routing_policy.requires_extension_parity);
-        assert!(command.required_extensions.is_empty());
-        assert!(!command.routing_policy.infer_source_path_tools);
-    }
-
-    #[test]
-    fn agent_task_controller_run_from_spec_supports_lab_placement_runner_routing() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "homeboy-lab",
-            "--placement",
-            "lab",
-            "agent-task",
-            "controller",
-            "run-from-spec",
-            "loop.json",
-            "--max-actions",
-            "1",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-        assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-        assert_eq!(cli.placement, crate::cli_surface::Placement::Lab);
-        assert_eq!(
-            command.hot_label,
-            "agent-task controller from-spec --resume/run-from-spec/materialize"
-        );
-        assert!(command.is_portable());
-        assert!(command.routing_policy.default_lab_offload);
-        assert!(!command.routing_policy.requires_extension_parity);
-        assert_eq!(
-            command.workspace_mode_policy,
-            runners::LabOffloadWorkspaceModePolicy::GitCheckoutRequired
-        );
-    }
-
-    #[test]
-    fn agent_task_controller_materialization_family_auto_selects_default_lab_runner() {
-        for args in [
-            [
-                "homeboy",
-                "agent-task",
-                "controller",
-                "from-spec",
-                "loop.json",
-                "--resume",
-                "--max-actions",
-                "1",
-            ]
-            .as_slice(),
-            [
-                "homeboy",
-                "agent-task",
-                "controller",
-                "run-from-spec",
-                "loop.json",
-                "--max-actions",
-                "1",
-            ]
-            .as_slice(),
-            [
-                "homeboy",
-                "agent-task",
-                "controller",
-                "materialize",
-                "loop.json",
-            ]
-            .as_slice(),
-        ] {
-            let cli = Cli::parse_from(args);
-
-            let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-            assert_eq!(
-                command.hot_label,
-                "agent-task controller from-spec --resume/run-from-spec/materialize"
-            );
-            assert!(command.is_portable());
-            assert!(command.routing_policy.default_lab_offload);
-            assert!(command.routing_policy.infer_source_path_tools);
-            assert!(!command.routing_policy.requires_extension_parity);
-            assert_eq!(
-                command.workspace_mode_policy,
-                runners::LabOffloadWorkspaceModePolicy::GitCheckoutRequired
-            );
-        }
-    }
-
-    #[test]
-    fn agent_task_fanout_submit_batch_requires_explicit_runner_under_lab_placement() {
-        // Isolate from a parallel test leaking the offload-metadata env var,
-        // which would otherwise short-circuit route_after_parse as a Lab
-        // offload subprocess and return Ok(None) instead of the deny error.
-        let _env = EnvGuard::remove(homeboy::core::observation::LAB_OFFLOAD_METADATA_ENV);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "--placement".to_string(),
-            "lab".to_string(),
-            "agent-task".to_string(),
-            "fanout".to_string(),
-            "submit-batch".to_string(),
-            "--input".to_string(),
-            "fanout.json".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-        assert_eq!(command.hot_label, "agent-task fanout submit-batch");
-        assert!(!command.routing_policy.default_lab_offload);
-        assert!(!command.routing_policy.infer_source_path_tools);
-
-        let err = route_after_parse(&cli, &normalized, None)
-            .expect_err("fanout submit-batch must not run locally under Lab placement");
-
-        assert_eq!(err.code.as_str(), "validation.invalid_argument");
-        // submit-batch needs an explicit runner under Lab placement: it does
-        // not auto-offload, so `--placement lab` without an eligible runner is
-        // refused rather than silently running locally.
-        assert!(err.message.contains("--placement lab"));
-        assert!(err.message.contains("requires an eligible Lab runner"));
-        assert!(err.message.contains("agent-task fanout submit-batch"));
-    }
-
-    #[test]
-    fn agent_task_fanout_run_plan_coordination_is_controller_local() {
-        let normalized = vec![
-            "homeboy".to_string(),
-            "--runner".to_string(),
-            "homeboy-lab".to_string(),
-            "--placement".to_string(),
-            "lab".to_string(),
-            "agent-task".to_string(),
-            "fanout".to_string(),
-            "run-plan".to_string(),
-            "--input".to_string(),
-            "fanout.json".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        // Fanout coordination is controller-owned so durable batch state,
-        // worktree ownership, and recovery stay local; the coordinator itself
-        // is not Lab-portable, though the independent cooks it generates may use
-        // their own Lab placement (#8045).
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-        assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-        assert_eq!(cli.placement, crate::cli_surface::Placement::Lab);
-        assert_eq!(command.hot_label, "agent-task fanout run-plan");
-        assert!(!command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-        assert!(!command.routing_policy.requires_extension_parity);
-        assert!(cli
-            .command
-            .lab_runner_unsupported_reason()
-            .is_some_and(|reason| reason.contains("controller-owned")));
-    }
-
-    #[test]
-    fn agent_task_fanout_cook_batch_run_plan_keeps_cook_coordinators_local() {
-        let normalized = vec![
-            "homeboy".to_string(),
-            "--runner".to_string(),
-            "homeboy-lab".to_string(),
-            "--placement".to_string(),
-            "lab".to_string(),
-            "agent-task".to_string(),
-            "fanout".to_string(),
-            "cook-batch".to_string(),
-            "--repo".to_string(),
-            "homeboy".to_string(),
-            "--verify".to_string(),
-            "cargo test --locked agent_task".to_string(),
-            "--run-plan".to_string(),
-            "https://github.com/Extra-Chill/homeboy/issues/7011".to_string(),
-        ];
-        let cli = Cli::parse_from(&normalized);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-        assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-        assert_eq!(cli.placement, crate::cli_surface::Placement::Lab);
-        assert_eq!(command.hot_label, "agent-task fanout cook-batch");
-        assert!(!command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-    }
-
-    #[test]
-    fn agent_task_fanout_state_reads_are_runner_resident() {
-        for args in [
-            [
-                "homeboy",
-                "--runner",
-                "homeboy-lab",
-                "agent-task",
-                "fanout",
-                "status",
-                "fanout-batch-123",
-            ],
-            [
-                "homeboy",
-                "--runner",
-                "homeboy-lab",
-                "agent-task",
-                "fanout",
-                "artifacts",
-                "fanout-batch-123",
-            ],
-        ] {
-            let cli = Cli::parse_from(args);
-
-            let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-            assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-            assert_eq!(command.hot_label, "agent-task fanout status/artifacts");
-            assert!(command.is_portable());
-            assert!(!command.routing_policy.default_lab_offload);
-            assert_eq!(
-                command.source_path_mode,
-                runners::LabOffloadSourcePathMode::RunnerResident
-            );
-            assert_eq!(
-                command.workspace_mode_policy,
-                runners::LabOffloadWorkspaceModePolicy::RunnerResident
-            );
-            assert!(command.required_extensions.is_empty());
-            assert!(!command.routing_policy.requires_extension_parity);
-            assert!(!command.routing_policy.infer_source_path_tools);
-        }
-    }
-
-    #[test]
-    fn tunnel_service_start_supports_explicit_runner_discovery() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "homeboy-lab",
-            "tunnel",
-            "service",
-            "start",
-            "preview",
-            "--cwd",
-            "/home/user/Developer/_lab_workspaces/site",
-            "--command",
-            "npm run dev",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(cli.runner.as_deref(), Some("homeboy-lab"));
-        assert_eq!(command.hot_label, "tunnel service start");
-        assert!(command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-        assert_eq!(
-            command.source_path_mode,
-            runners::LabOffloadSourcePathMode::RunnerResident
-        );
-        assert_eq!(
-            command.workspace_mode_policy,
-            runners::LabOffloadWorkspaceModePolicy::RunnerResident
-        );
-        assert!(!command.routing_policy.requires_extension_parity);
-        assert!(command.required_extensions.is_empty());
-        assert!(!command.routing_policy.infer_source_path_tools);
-    }
-
-    #[test]
-    fn tunnel_preview_consumer_run_keeps_explicit_runner_contract() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "--runner",
-            "homeboy-lab",
-            "tunnel",
-            "preview-consumer",
-            "run",
-            "--config",
-            "consumer.json",
-            "--preview-public-url",
-            "https://preview.example.test/run",
-        ]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "tunnel preview-consumer run");
-        assert!(command.is_portable());
-        assert!(!command.routing_policy.default_lab_offload);
-    }
-
-    #[test]
-    fn lab_command_with_mutation_flag_stays_portable_for_patch_capture() {
-        let cli = Cli::parse_from(["homeboy", "review", "audit", "--baseline"]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "review audit");
-        assert!(command.is_portable());
-        assert!(command.routing_policy.requires_extension_parity);
-    }
-
-    #[test]
-    fn lab_command_with_ratchet_stays_portable_for_patch_capture() {
-        let cli = Cli::parse_from(["homeboy", "review", "audit", "--ratchet"]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "review audit");
-        assert!(command.is_portable());
-        assert!(command.routing_policy.requires_extension_parity);
-    }
-
-    #[test]
-    fn lab_command_preserves_local_only_contract_shape() {
-        let cli = Cli::parse_from(["homeboy", "rig", "up", "demo"]);
-
-        let command = lab_offload_command(&cli.command).unwrap().unwrap();
-
-        assert_eq!(command.hot_label, "rig up");
-        assert!(matches!(
-            command.portability,
-            LabCommandPortability::LocalOnly(_)
-        ));
-        assert!(!command.routing_policy.requires_extension_parity);
-    }
-
-    #[test]
-    fn strip_component_target_replaces_positional_with_path() {
-        let args = vec![
-            "homeboy".to_string(),
-            "review".to_string(),
-            "lint".to_string(),
-            "--fix".to_string(),
-            "sample-component".to_string(),
-        ];
-
-        let rewritten = strip_component_target_args(&args, "sample-component", "/src/sample");
-
-        assert_eq!(
-            rewritten,
-            vec![
-                "homeboy".to_string(),
-                "review".to_string(),
-                "lint".to_string(),
-                "--fix".to_string(),
-                "--path".to_string(),
-                "/src/sample".to_string(),
-            ]
-        );
-    }
-
-    #[test]
-    fn strip_component_target_replaces_component_flag_with_path() {
-        let args = vec![
-            "homeboy".to_string(),
-            "refactor".to_string(),
-            "--from".to_string(),
-            "lint".to_string(),
-            "--write".to_string(),
-            "--component".to_string(),
-            "sample-component".to_string(),
-        ];
-
-        let rewritten = strip_component_target_args(&args, "sample-component", "/src/sample");
-
-        assert_eq!(
-            rewritten,
-            vec![
-                "homeboy".to_string(),
-                "refactor".to_string(),
-                "--from".to_string(),
-                "lint".to_string(),
-                "--write".to_string(),
-                "--path".to_string(),
-                "/src/sample".to_string(),
-            ]
-        );
-    }
-
-    #[test]
-    fn strip_component_target_only_strips_first_positional_match() {
-        // A `--from` value equal to the component id must survive; only the bare
-        // positional component token is dropped.
-        let args = vec![
-            "homeboy".to_string(),
-            "refactor".to_string(),
-            "--from".to_string(),
-            "lint".to_string(),
-            "--write".to_string(),
-            "dmc".to_string(),
-        ];
-
-        let rewritten = strip_component_target_args(&args, "dmc", "/src/dmc");
-
-        assert_eq!(
-            rewritten,
-            vec![
-                "homeboy".to_string(),
-                "refactor".to_string(),
-                "--from".to_string(),
-                "lint".to_string(),
-                "--write".to_string(),
-                "--path".to_string(),
-                "/src/dmc".to_string(),
-            ]
-        );
-    }
-
-    #[test]
-    fn strip_component_target_preserves_passthrough_args() {
-        let args = vec![
-            "homeboy".to_string(),
-            "lint".to_string(),
-            "--fix".to_string(),
-            "dmc".to_string(),
-            "--".to_string(),
-            "dmc".to_string(),
-        ];
-
-        let rewritten = strip_component_target_args(&args, "dmc", "/src/dmc");
-
-        assert_eq!(
-            rewritten,
-            vec![
-                "homeboy".to_string(),
-                "lint".to_string(),
-                "--fix".to_string(),
-                "--".to_string(),
-                "dmc".to_string(),
-                "--path".to_string(),
-                "/src/dmc".to_string(),
-            ]
-        );
-    }
-
-    #[test]
-    fn rewrite_component_target_skips_when_path_override_present() {
-        let cli = Cli::parse_from([
-            "homeboy",
-            "review",
-            "lint",
-            "--fix",
-            "sample-component",
-            "--path",
-            "/explicit/path",
-        ]);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "review".to_string(),
-            "lint".to_string(),
-            "--fix".to_string(),
-            "sample-component".to_string(),
-            "--path".to_string(),
-            "/explicit/path".to_string(),
-        ];
-
-        assert!(rewrite_component_target_to_path(&cli.command, &normalized).is_none());
-    }
-
-    #[test]
-    fn rewrite_component_target_skips_without_component() {
-        // No positional component and no --path: source resolves from CWD, so
-        // there is nothing to rewrite.
-        let cli = Cli::parse_from(["homeboy", "review", "lint", "--fix"]);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "review".to_string(),
-            "lint".to_string(),
-            "--fix".to_string(),
-        ];
-
-        assert!(rewrite_component_target_to_path(&cli.command, &normalized).is_none());
-    }
-
-    #[test]
-    fn lab_route_source_path_args_rewrites_review_lint_component_without_patch_capture() {
-        let cli = Cli::parse_from(["homeboy", "review", "lint", "homeboy"]);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "review".to_string(),
-            "lint".to_string(),
-            "homeboy".to_string(),
-        ];
-
-        let rewritten = lab_route_source_path_args(&cli.command, &normalized, false)
-            .expect("review lint component id should become a source path");
-
-        assert_eq!(rewritten[0..3], normalized[0..3]);
-        assert_eq!(
-            rewritten
-                .iter()
-                .filter(|arg| arg.as_str() == "homeboy")
-                .count(),
-            1
-        );
-        assert!(rewritten.contains(&"--path".to_string()));
-    }
-
-    #[test]
-    fn rewrite_ad_hoc_lab_workspace_adds_path_for_pathless_lint() {
-        let dir = tempdir().unwrap();
-        let _cwd = CwdGuard::set(dir.path());
-        let cwd = std::env::current_dir().expect("current dir");
-        let cli = Cli::parse_from(["homeboy", "review", "lint"]);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "review".to_string(),
-            "lint".to_string(),
-        ];
-
-        let rewritten = rewrite_ad_hoc_lab_workspace_to_path(&cli.command, &normalized)
-            .expect("pathless lint should become explicit path");
-
-        assert_eq!(
-            rewritten,
-            vec![
-                "homeboy".to_string(),
-                "review".to_string(),
-                "lint".to_string(),
-                "--path".to_string(),
-                cwd.to_string_lossy().to_string(),
-            ]
-        );
-    }
-
-    #[test]
-    fn rewrite_ad_hoc_lab_workspace_inserts_path_before_passthrough() {
-        let dir = tempdir().unwrap();
-        let _cwd = CwdGuard::set(dir.path());
-        let cwd = std::env::current_dir().expect("current dir");
-        let cli = Cli::parse_from(["homeboy", "review", "test", "--", "--filter", "ExampleTest"]);
-        let normalized = vec![
-            "homeboy".to_string(),
-            "review".to_string(),
-            "test".to_string(),
-            "--".to_string(),
-            "--filter".to_string(),
-            "ExampleTest".to_string(),
-        ];
-
-        let rewritten = rewrite_ad_hoc_lab_workspace_to_path(&cli.command, &normalized)
-            .expect("pathless test should become explicit path");
-
-        assert_eq!(
-            rewritten,
-            vec![
-                "homeboy".to_string(),
-                "review".to_string(),
-                "test".to_string(),
-                "--path".to_string(),
-                cwd.to_string_lossy().to_string(),
-                "--".to_string(),
-                "--filter".to_string(),
-                "ExampleTest".to_string(),
-            ]
-        );
-    }
-
-    #[test]
-    fn rewrite_ad_hoc_lab_workspace_skips_registered_component_or_path() {
-        let component_cli = Cli::parse_from(["homeboy", "review", "lint", "homeboy"]);
-        let path_cli = Cli::parse_from(["homeboy", "review", "audit", "--path", "/tmp/homeboy"]);
-
-        assert!(rewrite_ad_hoc_lab_workspace_to_path(
-            &component_cli.command,
-            &[
-                "homeboy".to_string(),
-                "review".to_string(),
-                "lint".to_string(),
-                "homeboy".to_string(),
-            ],
-        )
-        .is_none());
-        assert!(rewrite_ad_hoc_lab_workspace_to_path(
-            &path_cli.command,
-            &[
-                "homeboy".to_string(),
-                "review".to_string(),
-                "audit".to_string(),
-                "--path".to_string(),
-                "/tmp/homeboy".to_string(),
-            ],
-        )
-        .is_none());
-    }
-}
+mod tests;
